requirements.txt
wxpython
requests
pandas
matplotlib
openpyxl
python-docx
pypdf
